+++ src/App.js
@@ -1,17 +1,20 @@
-import { Box } from "@mui/material";
 import "./App.css";
 import { Login } from "./components/auth/Login";
-import { Register } from "./components/auth/Register";
 import Home from "./components/Home";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Landing from "./components/Landing";
+
 
 function App() {
   return (
     <>
-      <Box sx={{ width: "100%", m: 2 }}>
-        <Home />
-        <Login />
-        <Register/>
-      </Box>
+      <BrowserRouter>
+      <Landing/>
+      <Routes>
+        <Route path={"/"} element={<Home/>}/>
+        <Route path={"/Login"} element={<Login/>}/>
+      </Routes>
+      </BrowserRouter>
     </>
   );
 }
