+++ src/components/utils/constants.js
@@ -1,0 +1,6 @@
+export const navHeaders = [
+    { id: "1", NavValue: "Home", url: "/" },
+    { id: "2", NavValue: "Add New Button", url: "/AddNew" },
+    { id: "3", NavValue: "Login", url: "/Login" },
+  ];
+
